(EAI-598) [UI] Investigate further code splitting (#563)
