+++ packages/mongodb-chatbot-ui/src/ChatbotView.tsx
@@ -1,16 +1,15 @@
 import { References } from "mongodb-rag-core";
+import { ChatMessageFeedProps } from "./ChatMessageFeed";
 import { DarkModeProps } from "./DarkMode";
 
-export type ChatbotViewProps = DarkModeProps & {
-  className?: HTMLElement["className"];
-  disclaimer?: React.ReactNode;
-  disclaimerHeading?: string;
-  fatalErrorMessage?: string;
-  initialMessageReferences?: References;
-  initialMessageSuggestedPrompts?: string[];
-  initialMessageText?: string;
-  inputBarId?: string;
-  inputBarPlaceholder?: string;
-  inputBottomText?: string;
-  windowTitle?: string;
-};
+export type ChatbotViewProps = DarkModeProps &
+  Omit<ChatMessageFeedProps, "initialMessage"> & {
+    fatalErrorMessage?: string;
+    initialMessageReferences?: References;
+    initialMessageSuggestedPrompts?: string[];
+    initialMessageText?: string;
+    inputBarId?: string;
+    inputBarPlaceholder?: string;
+    inputBottomText?: string;
+    windowTitle?: string;
+  };
